+++ src/TgBotLib/Abstracts/StickerFormat.php
@@ -1,0 +1,10 @@
+<?php
+
+    namespace TgBotLib\Abstracts;
+
+    abstract class StickerFormat
+    {
+        const Static = 'static';
+        const Animated = 'animated';
+        const Video = 'video';
+    }